Refactor timer GUI into its own module (actually MVC this time?)
